fix(member-service): PaymenteEventConsumer endpoint

+++ member-service/src/main/java/com/gym/member_service/models/PaymentRecord.java
@@ -16,5 +16,5 @@ public class PaymentRecord {
     private Double amount;
     private LocalDateTime paymentDate;
     private String promotionId;
-    private Double discountPercentage;
+    private Long discountPercentage;
 }
